moar refactor, stripped the classes out of their component and put into their own files
