+++ src/app/classes/dimensions.ts
@@ -1,0 +1,9 @@
+export class Dimensions {
+    x: number;
+    y: number;
+
+    constructor(x?: number, y?: number) {
+        this.x = x;
+        this.y = y;
+    }
+}
